Renamed the generated prettyasserts's extension to .a.c
So instead of preprocessing lfs.t.a.c -> lfs.t.c, we preprocess
lfs.t.c -> lfs.t.a.c.

Really the extension should indicate what tool it was generated by, not
what tool it should be consumed by.

The previous commit that changed this states:

> Changed prettyasserts.py rule to .a.c => .c, allowing other .a.c files
> in the future.

But I'm not really sure why we would ever not just run prettyasserts.py
on every C file...

External use of prettyasserts.py made it clear the previous naming was a
bit weird.