Changed all test/bench runner build-time features to opt-in
- NO_COV    -> YES_COV
- NO_PERF   -> YES_PERF
- NO_PERFBD -> YES_PERFBD

Previously, COV defaulted to yes for tests, and PERFBD defaulted to yes
for benches. This is sometimes useful, but much less often than I
originally thought. Might as well not pay for what we don't use.

With this, the build features of the test/bench runners are consistent
by default, which is probably a good thing.

This _does_ have a noticable, if minor, impact on test runtime:

  YES_COV: 674.15s
  NO_COV:  584.97s (-13.2%)

As for the naming, the YES_* prefix is needed to avoid conflicts with
the tool variables themselves. I'm not sure what the best approach to
variable naming is here...

  $ YES_PERF=1 PERF=~/my_perf/my_perf make test-runner -j