Changed pcache/rcache interactions to wait to overwrite until flush
Previous versions of littlefs saw very little pcache/rcache interaction,
which was a nice simplification for the bd layer. But now, with rbyds,
we rely overlapping pcaches/rcaches heavily. This is because building
each rbyd trunk requires reading the previous rbyd trunk, which may have
not made it to disk yet.

The main issue this presents, is that reads always need to prioritize
data in the pcache, even if it doesn't exist on disk yet.

This gets a bit annoying with read/prog alignment requirements, which
may require disk-reads that overlap the pcache.

And even more annoying when you consider that after a flush, the rcache
should reflect the new data even if pcache is dropped.

The fact that the current impl works at all is because of tests and
sweat...

---

To solve these problems, the bd layer would overwrite the rcache on
prog. This alone wasn't sufficient however, as we also need to overwrite
the rcache on reads because of the above alignment issue.

So:
               pcache            rcache
               ................  ................
  read(0..4)   ................  aaaa............
  prog(6..10)  ......bbbb......  aaaa..bbbb......
  read(0..8)   ......bbbb......  aaaaccbbbb...... => aaaaccbb
  flush()      ................  aaaaccbbbb......
  read(0..8)   ................  aaaaccbbbb...... => aaaacbbb

Note we can't just not overwrite the rcache, since flushing the pcache
leaves us with out-of-date information:

               pcache            rcache
               ................  ................
  read(0..4)   ................  aaaa............
  prog(6..10)  ......bbbb......  aaaa............
  read(0..8)   ......bbbb......  aaaacccc........ => aaaaccbb
  flush()      ................  aaaacccc........
  read(0..8)   ................  aaaacccc........ => aaaacccc !!!

This commit adopts a slightly different strategy: overwrite when we
flush:

               pcache            rcache
               ................  ................
  read(0..4)   ................  aaaa............
  prog(6..10)  ......bbbb......  aaaa............
  read(0..8)   ......bbbb......  aaaacccc........ => aaaaccbb
  flush()      ................  aaaaccbbbb......
  read(0..8)   ................  aaaaccbbbb...... => aaaaccbb

This keeps the rcache always in sync with disk (we don't care if pcache
is dropped without a flush), leaving unflushed pcache overwrites up to
lfsr_bd_read, which it needs to handle correctly anyways because of the
above alingment issue.

This saves a single overwrite.

Which isn't really that much when it comes to code cost:

           code          stack
  before: 33560           2808
  after:  33544 (-0.0%)   2800 (-0.3%)

But hey at least we're doing fewer copies? And no one should be tempted
to remove the overwrite-on-read code thinking it's redundant now (wasn't
me!).