Changed lfsr_bd_set to hijack the rcache
This is a compromise between using a small hardcoded buffer and
cache-access during progs. Instead of getting direct access to the
pcache during progs, we just hijack the rcache, forcefully evicting any
contents it might have.

This gets us cache-access (of at least some cache) without needing to
rewrite lfsr_bd_prog.

The downside is this may result in more rcache misses. Though the use of
lfsr_bd_set is fairly niche in littlefs, so hopefully this doesn't
become a problem.

Code changes:

           code          stack
  before: 33796           2880
  after:  33792 (-0.0%)   2872 (-0.3%)