Changed DATA macros to implicitly stack allocate via compound literals
So instead of:

  uint8_t mptr_buf[LFSR_MPTR_DSIZE];
  int err = lfsr_btree_commit(lfs, &mtree_.u.btree, 0, LFSR_ATTRS(
          LFSR_ATTR(
              MDIR, +lfsr_mleafweight(lfs),
              FROMMPTR(lfsr_mdir_mptr(&mdir_), &mptr_buf))));

This can be written as:

  int err = lfsr_btree_commit(lfs, &mtree_.u.btree, 0, LFSR_ATTRS(
          LFSR_ATTR(
              MDIR, +lfsr_mleafweight(lfs),
              FROMMPTR(lfsr_mdir_mptr(&mdir_)))));

Explicit stack allocation is still possible with the DATA hole, though a
bit more annoying:

  attrs[attr_count++] = LFSR_ATTR(
          MDIR, +lfsr_mleafweight(lfs),
          DATA(lfsr_data_frommptr(
              lfsr_mdir_mptr(&mdir_),
              &buf[buf_size])));
  buf_size += LFSR_MPTR_DSIZE;

The main motivation for this change is to be consistent with
LFSR_DATA_CAT, which was already implicitly stack allocating. The macros
that take arrays are relatively error-prone otherwise (LFSR_DATA_CAT,
LFSR_ATTRS, etc).

This does come with the benefit that the required buffer size is
implicitly provided by the macro, so no worry of it falling out-of-sync
externally. However this does come with the tradeoff of compound literal
lifetimes, which requires the result to live only as long as the current
expression.

Hopefully the fact that these are MACROs signal that they need special
care to any new developers...

Unfortunately, the use of compound literals also brings a surprising
code/stack cost:

           code          stack
  before: 33912           2872
  after:  34016 (+0.3%)   2896 (+0.8%)

Currently I can think of two reasons:

1. It's not possible to declare an uninitialized compound literal.

   This probably sounds like a good thing to memory-safety fans, and
   initialized is probably a good default for variable declaration, but
   the reality is the required initialization does add useless code.

   This specific use of compound literals is also low-risk given that we
   immediately pass the literal to an lfsr_data_from* function, which
   does the initialization.

2. We sometimes share on-stack buffers between branches of ternary
   expressions since we know their use is exclusive. These macros sort
   of get in the way of that.

What I find a bit curious is GCC doesn't seem capable of optimizating
away these overheads, which I would think would be possible given that
GCC knows all the information of how these buffers end up used.

I've noticed in general compound literals add overhead when the
underlying semantics don't really change. I wonder if this is because
compound literals are relatively new/unused, or some required
side-effects I'm missing. Maybe this will improve in the future?

Anyways, I'm keeping this change for now, since it does improve the
internal attr-list ergonomics/safety. Though these sort of changes are
always open to be revisited in the future.

Interestingly, the future-theoretical transpilation to c89 may save
code/stack because of this, which raises some questions...