Rearranged lfsr_file_carve a bit to avoid attr_tnuoc shenanigans
The reason for carving up the right sibling before appending our new
data is because we 1. want to carve both left+right siblings in a single
lookup if possible, and 2. we don't want to keep unnecessary lookup state
around as much as possible.

But is keeping some lookup state around cheaper than the attr_tnuoc +
memmove mess? The answer is yes:

           code          stack
  before: 34028           2896
  after:  33944 (-0.2%)   2888 (-0.3%)

attr_tnuoc is one of those "if it's stupid and it works it's not stupid"
solutions, but that doesn't make it not stupid. (I'm joking a bit, but
the new code is cleaner + more readable, which was the original
motivation for looking at this function again)