Improved rbyd compaction algorithm, reduced attr estimate 3t+4 -> (5/2)t+2
The motivation for this is that the rbyd inner node encoding during
compaction is kind-of not that great.

Our alt encoding is great when the trunk terminates in a tag, which is
how it was originally designed to be used:

  00000004: data w1 1                61           a     <.
  00000020: altble 0x300 w1 0x4                         -' <-- trunk
  00000024: data w1 1                62           b

But when used to create an arbitrary binary-tree inner node, the best
encoding I can think of is 2 alts + a terminating null tag, which is not
that great:

  00000004: data w1 1                61           a     <.
  00000009: data w1 1                62           b     <--.
  0000000e: altble 0x300 w1 0x4                         -' | <-- trunk
  00000012: altble 0x300 w1 0x9                         ---'
  00000016: null

This effects our attr estimate, which is defined as the worst-case
on-disk cost of an attr after compaction. This is an important value,
as it determines when we split rbyds. And it effectively determines how
densely we can store metadata without needing to worry about block
overflow issues.

In our current compaction algorithm, we connect each attr with a 2 alt +
null inner node. Since we are creating a perfectly balanced binary tree,
this works out to ~1 inner node per attr. Including the attr's data tag,
this gives us:

  a_0 = 3t + 4

Where t is the tag estimate, currently a 2 byte tag, <=5 byte weight,
<=4 byte size, t = 2+5+4 = 11 bytes:

  a_0 = 3*11 + 4 = 37 bytes

Though this may be vary across different littlefs configurations,
16-bit, 64-bit, etc.

---

It would be great if our compaction algorithm could build each trunk
perfectly, as each attr is written. In such a case, each attr
theoretically only needs ~1 alt and the attr's data tag:

  a_inf = 2t

Or, assuming t = 11 bytes:

  a_inf = 2*11 = 22 bytes

Unfortunately, as far as I can tell, this fundamentally requires
unbounded RAM. You need to keep track of log n previous trunks in order
to always build the next trunk perfectly, and log n is > 1.

I suppose in theory you could implement a O(n^2) algorithm that
repeatedly scans for the previous trunks... But that would be a
hilarious regression since the whole point of this work was to reduce
compaction from O(n^2) -> O(log n).

---

However, we can meet halfway. Consider what happens if we build perfect
trunks for only the bottom layer of the rbyd.

This may not seem like it will gain much, but remember that in a binary
tree, the bottom layer contains ~1/2 of the total nodes in the tree:

        3t + 4   2t   5t
  a_1 = ------ + -- = -- + 2
           2      2    2

Or, assuming t = 11 bytes:

        5*11
  a_1 = ---- + 2 = 30 bytes
          2

Not too shabby for a constant amount of RAM.

In theory this could be extended to n layers, by keeping a
(configurable?) array of previous trunks in RAM during compaction, but
this would have diminishing results as we move up the tree.

With only needing to keep track of one other trunk, we can even store
this in rybd.trunk, which is currently unused during compaction. So zero
extra RAM.

The resulting compaction looks like the following:

  00000004: data w1 1                61           a     <.
  00000020: altble 0x300 w1 0x4                         <--.
  00000024: data w1 1                62           b        |
  00000029: data w1 1                63           c     <. |
  0000002e: altble 0x300 w1 0x29                        <. |
  00000032: data w1 1                64           d      | |
  00000037: altble 0x300 w2 0x20                        -' | <-- trunk
  0000003b: altble 0x300 w2 0x2e                        ---'
  0000003f: null

This does make compaction a bit more complicated, which is reflected in
the code size:

           code          stack
  before: 33856           2880
  after:  34068 (+0.6%)   2880 (+0.0%)

Some implementation things to note:

- Since lfsr_rbyd_appendcompactattr now has state, we need to make sure
  to zero the trunk before compacting, which adds an annoying bit of
  bookkeeping everywhere.

- I didn't want to add extra state to manage in
  lfsr_rbyd_appendcompactattr calls, so this implementation only tracks
  the previous trunk offset, and needs a readtag call to get the
  tag+weight necessary to build the actual alt pointer.

  This may have some unpleasant interactions with the rcache, and may be
  worth revisiting.