Added size-limit to our runtime dependent attr estimate
This shouldn't impact most systems, but it's not unreasonable to allow
size-limit to be tweaked for the sole purpose of better metadata
density.

As a plus, if we mount a smaller filesystem, say a 16-bit littlefs, we'd
naturaly inherit its attr estimate/metadata density. This is probably
the more important side-effect.

This math is complicated enough to come with a bit of a code cost:

           code          stack
  before: 34104           2880
  after:  34132 (+0.1%)   2880 (+0.0%)

Though if we move size-limit + block-size to compile-time in the future,
this actually becomes free.