Opportunistically stitch together range removals with red alts
This is intending to improve some balancing issues with range removals
in our rbyds.

Consider the following range removal, this is our current algorithm:

                                          .-----------o
         .-------o-------.                |           o-----------.
     .---o---.       .---o---.            |     .-----o           |
   .-o-.   .-o-.   .-o-.   .-o-.        .-o-.   |     o-----.   .-o-.
  .o. .o. .o. .o. .o. .o. .o. .o.      .o. .o. .o. .--o--. .o. .o. .o.
  a b c d e f g h i j k l m n o p  =>  a b c d e f g     j k l m n o p
               '-+-'
               remove

Somehow the height of the tree increased! Even though we are only
removing nodes. Not great.

The reason this happens is because we are trying to stitch together the
two search paths that occur when our range diverges. Naively, with
binary nodes, this results in a worst case of ~2x the diverged height.

If only there was a way to represent a ternary node... Wait, isn't this
what our red alts are for?

Recall that in a red-black(-yellow) tree, red edges are a coloring that
represent a 2-3-4 node with 3 branches. If, as we stitch together our
two search paths, we alternate between red and black alts, we can avoid
a height increase in the underlying 2-3-4 tree!

         .-------o-------.
     .---o---.       .---o---.            .-----------r-----------.
   .-o-.   .-o-.   .-o-.   .-o-.        .-o-.   .-----r-----.   .-o-.
  .o. .o. .o. .o. .o. .o. .o. .o.      .o. .o. .o. .--o--. .o. .o. .o.
  a b c d e f g h i j k l m n o p  =>  a b c d e f g     j k l m n o p
               '-+-'
               remove

This works great if all our nodes are black. Unfortunately, if we
already have red alts, this doesn't always work. We can't connect red
alts with red alts, or we risk breaking invariants:

                                    .---+-------r
       .-------o-------.            |   |       r-------+---.
   .---r----.     .----r---.        |   |  .-+--r       |   |
  .o. .o. .-r-. .-r-. .o. .o.      .o. .o. | |  o--+-. .o. .o.
  a b c d e f g h i j k l m n  =>  a b c d e f     i j k l m n
             '-+-'
             remove

If we ignore red alts, and pretend they are black during range removals,
we just end up with a slightly permuted tree:

                                    .-----------r-----------.
       .-------o-------.            |   .-------r-------.   |
   .---r----.     .----r---.        |   |  .----r----.  |   |
  .o. .o. .-r-. .-r-. .o. .o.      .o. .o. | .--o--. | .o. .o.
  a b c d e f g h i j k l m n  =>  a b c d e f     i j k l m n
             '-+-'
             remove

It's tempting to try to stitch red nodes together with yellow alts, but
this breaks the invariant that the our parent is never yellow during
append, forcing append to be potentially recursive if we encounter
a naturally occuring yellow alt.

With a hypothetical 5-branch node however...

But at least this delays unbalancing when black alts are present.

And, since we downgrade any red alts when pruning during range removals,
we should end up with more black alts available for opportunistic
stitching than in the original tree.

---

Surprisingly the code cost ended up breaking even, probably because of
some minor code cleanup in the function:

  before: 34072          2880
  after:  34072 (+0.0%)  2880 (+0.0%)

Measuring performance with a quick file random-write benchmark showed
a noticable but tiny improvement. Though it may be 1. too close of to
the noise floor to be trustworthy, 2. not really rigorous, and 3. file
random-write may not hit dgenerate range removals. But hey, at least it
doesn't show a negative impact on performance.